Add files via upload
Using the BMI scale, this is a simple interactive Octave/Matlab program I created that determines if the user-inputted weight/height combination is healthy according to the BMI index. If it is either underweight or overweight, the program tells the user how many pounds underweight or overweight the inputted weight is according to the inputted height.